requirements.txt
py-xgboost==1.6.1
pandas==1.4.3
bs4==4.11.1
requests==2.28.1
openpyxl==3.0.10
